+++ Changelog.md
@@ -1,3 +1,10 @@
+## Gazebo 9
+
+## Gazebo 9.x.x (2018-xx-xx)
+
+1. Remove Gazebo 8 deprecations
+    * [Pull request #2605](https://bitbucket.org/osrf/gazebo/pull-request/2605)
+
 ## Gazebo 8
 
 ## Gazebo 8.x.x (2017-xx-xx)